feat: Update volumetric rendering to use UINT textures for atomic operations
- Changed the volume texture format from R16_FLOAT to R32_UINT in the volumetric rendering shaders to support atomic operations, preventing race conditions at high particle counts.
- Implemented diagnostic counters to track shader execution metrics, including total threads executed and voxel writes, enhancing debugging capabilities.
- Updated the `PopulateVolumeMip2` method to utilize atomic operations for voxel density accumulation, ensuring correct behavior under high load.
- Enhanced the `Application` class to read and log diagnostic counters after the first frame, providing insights into shader performance.

These changes improve the stability and performance of the volumetric rendering system, addressing GPU hang issues and facilitating better debugging.

+++ shaders/volumetric_restir/path_generation.hlsl
@@ -60,7 +60,8 @@ struct Particle {
 StructuredBuffer<Particle> g_particles : register(t1);
 
 // Piecewise-constant transmittance volume (Mip 2, coarse grid)
-Texture3D<float> g_volumeMip2 : register(t2);
+// IMPORTANT: Stored as UINT for atomic operations, convert to float with asfloat()
+Texture3D<uint> g_volumeMip2 : register(t2);
 SamplerState g_volumeSampler : register(s0);
 
 // Output: reservoir buffer (RWStructuredBuffer)
